+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - we're aware of Lastfm API Changes, we removed all of their APIs from Dawayer as a response, sorry for the inconvenience, we will have a solution soon.
 - support for opening a track file while Dawayer is already running.
 - fixed an error that appears when scanning a non-existing directories.
+- fixed a bug where shuffle always plays the first track instead of the random.